+++ src/recipe.h
@@ -22,6 +22,7 @@ namespace forge
     std::string url;
     std::string sha256;
     std::string github;
+    std::string package;
     std::string version;
     std::string git;
     std::string commit;